docshell/test/navigation/test_bug1380305.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1380305
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1380305</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1380305 **/
  SimpleTest.waitForExplicitFinish();

  function test() {
    let webNav = SpecialPowers.wrap(window.top)
                 .QueryInterface(SpecialPowers.Ci.nsIInterfaceRequestor)
                 .getInterface(SpecialPowers.Ci.nsIWebNavigation);
    let maxLength = webNav.sessionHistory.maxLength;
    ok(maxLength > 0, "sanity check of maxLength");

    // Push maxLength states in a short time.
    let ref = "ref";
    for (let i = 0; i < maxLength; i++) {
      history.pushState(null, "test", location.href + "#" + ref);
      ref += i;
    }

    // The next pushState should throw.
    SimpleTest.doesThrow(
      () => history.pushState(null, "test", location.href + "#" + ref),
      "check history.pushState does throw");

    SimpleTest.finish();
  }

  </script>
</head>
<body onload="setTimeout(test, 0);">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1380305">Mozilla Bug 1380305</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
